Split knit publish.rs: extract the PR-body block renderer
Move the managed "Knit Bundle" cross-link block (markers, initial_pr_body,
render/upsert/append) into publish/pr_body.rs. publish/mod.rs keeps the create
and sync entrypoints (live + from-artifact), repo-scope resolution, and
BaseOverrides. No behavior change; tests and clippy unchanged.


Knit-Group: kg_20260531_1833c5
Knit-Bundle: split-more-oversized-files-round-3
