+++ src/commands/publish/mod.rs
@@ -14,8 +14,8 @@ use std::collections::{BTreeMap, BTreeSet};
 use std::ffi::OsString;
 use std::path::Path;
 
-const KNIT_PR_BLOCK_BEGIN: &str = "<!-- BEGIN KNIT BUNDLE -->";
-const KNIT_PR_BLOCK_END: &str = "<!-- END KNIT BUNDLE -->";
+mod pr_body;
+use pr_body::{initial_pr_body, render_knit_pr_block, upsert_knit_pr_block};
 
 pub fn create_publications(
     selectors: &[String],
@@ -752,80 +752,6 @@ fn run_push(cwd: &Path, branch: &str, set_upstream: bool) -> Result<()> {
     Ok(())
 }
 
-fn initial_pr_body(bundle: &ChangeGroup, current_repo_id: &str) -> String {
-    format!(
-        "Generated by Knit. Cross-links will be refreshed after all review objects are created.\n\n{}",
-        render_knit_pr_block(bundle, Some(current_repo_id))
-    )
-}
-
-fn render_knit_pr_block(bundle: &ChangeGroup, current_repo_id: Option<&str>) -> String {
-    let mut lines = vec![
-        KNIT_PR_BLOCK_BEGIN.to_string(),
-        "## Knit Bundle".to_string(),
-        String::new(),
-        format!("This PR is part of Knit bundle `{}`.", bundle.id),
-        String::new(),
-        "See the other review objects in this bundle:".to_string(),
-    ];
-
-    let repo_ids = publish_scope_repo_ids(bundle);
-    let scoped_repos = bundle
-        .repos
-        .iter()
-        .filter(|repo| repo_ids.is_empty() || repo_ids.contains(&repo.id));
-
-    for repo in scoped_repos {
-        match publication_for_repo(bundle, &repo.id) {
-            Some(pr) => {
-                let marker = if current_repo_id == Some(repo.id.as_str()) {
-                    " (this PR)"
-                } else {
-                    ""
-                };
-                lines.push(format!("- `{}`: {}{}", repo.id, pr.url, marker));
-            }
-            None => lines.push(format!("- `{}`: pending", repo.id)),
-        }
-    }
-
-    lines.extend([
-        String::new(),
-        format!("Bundle id: `{}`", bundle.id),
-        format!("Bundle title: {}", bundle.title),
-        KNIT_PR_BLOCK_END.to_string(),
-    ]);
-
-    lines.join("\n")
-}
-
-fn upsert_knit_pr_block(existing_body: &str, block: &str) -> String {
-    let Some(begin) = existing_body.find(KNIT_PR_BLOCK_BEGIN) else {
-        return append_knit_pr_block(existing_body, block);
-    };
-    let Some(relative_end) = existing_body[begin..].find(KNIT_PR_BLOCK_END) else {
-        return append_knit_pr_block(existing_body, block);
-    };
-    let end = begin + relative_end + KNIT_PR_BLOCK_END.len();
-
-    let before = existing_body[..begin].trim_end();
-    let after = existing_body[end..].trim_start();
-    match (before.is_empty(), after.is_empty()) {
-        (true, true) => block.to_string(),
-        (true, false) => format!("{block}\n\n{after}"),
-        (false, true) => format!("{before}\n\n{block}"),
-        (false, false) => format!("{before}\n\n{block}\n\n{after}"),
-    }
-}
-
-fn append_knit_pr_block(existing_body: &str, block: &str) -> String {
-    if existing_body.trim().is_empty() {
-        block.to_string()
-    } else {
-        format!("{}\n\n{}", existing_body.trim_end(), block)
-    }
-}
-
 fn write_bundle_artifact_output(bundle: &ChangeGroup, out_path: Option<&Path>) -> Result<()> {
     match out_path {
         Some(path) => crate::store::write_json(path, bundle),
@@ -840,6 +766,7 @@ fn write_bundle_artifact_output(bundle: &ChangeGroup, out_path: Option<&Path>) -
 #[cfg(test)]
 mod tests {
     use super::*;
+    use super::pr_body::{KNIT_PR_BLOCK_BEGIN, KNIT_PR_BLOCK_END};
     use crate::model::{
         CommitGroup, CommitRef, PublicationEntry, CHANGE_GROUP_KIND, SCHEMA_VERSION,
     };
